refactor: chat-addon 구성 컴포넌트들 별도 파일 분리
