+++ apps/service/src/widgets/chat-addon/components/index.ts
@@ -1,0 +1,3 @@
+export { ChatRoomList } from "./chat-room-list";
+export { ChatRoomView } from "./chat-room-view";
+export { LoginPrompt } from "./login-prompt";
